Fix Listener name with suffx and add db transactions
