+++ app/Providers/EventServiceProvider.php
@@ -3,7 +3,7 @@
 namespace App\Providers;
 
 use App\Events\EventCreated;
-use App\Listeners\CreateSchedule;
+use App\Listeners\CreateScheduleListener;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
@@ -20,7 +20,7 @@ class EventServiceProvider extends ServiceProvider
             SendEmailVerificationNotification::class,
         ],
         EventCreated::class => [
-            CreateSchedule::class
+            CreateScheduleListener::class
         ],
     ];
